Add files via upload
I've been having issues with the database on my laptop so had to recreate it on a different PC. Hopefully this has the 3 tables that I specified in the whatsapp group.  It doesn't have any entries though. Let me know if it's fine and if there is anything I need to change about it.